backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
pillow==11.1.0
PyJWT==2.9.0
psycopg2-binary
django-environ
gunicorn